requirements.txt
# Lochross Outreach Bot - Dependencies
# Python 3.11+

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Google Sheets API
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Anthropic Claude API
anthropic>=0.18.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Utilities
lxml>=4.9.0
